+++ packages/driver-universal/src/dom.js
@@ -1,12 +1,13 @@
 import * as DriverDOM from 'driver-dom';
 
 export default function createDOMDriver() {
-  return Object.assign({}, DriverDOM, {
+  return {
+    ...DriverDOM,
     createElement(type, props, component) {
       return DriverDOM.createElement(type, props, component, true);
     },
     setStyle(node, style) {
       return DriverDOM.setStyle(node, style, true);
     }
-  });
+  };
 }