+++ packages/rax-pwa/src/index.js
@@ -1,8 +1,0 @@
-export Document from './Document';
-export Error from './Error';
-export TabBar from './TabBar/index';
-export Navigation from './Navigation/index';
-export preload from './preload';
-export prerender from './prerender';
-export { createRouter, router } from './createRouter';
-export getCurrentComponent from './getCurrentComponent';